requirements.txt
Django==1.4.1
Jinja2==2.6
Pygments==1.5
PyYAML==3.10
Sphinx==1.1.3
amqplib==1.0.2
anyjson==0.3.3
billiard==2.7.3.15
blinker==1.2
-e git+https://github.com/clement/brukva.git@bff451511a3cc09cd52bebcf6372a59d36567827#egg=brukva-0.0.1
celery==3.0.10
django-appconf==0.5
django-browserid==0.6
django-celery==3.0.10
django-classy-tags==0.3.4.1
django-missing==0.1.11
django-mongo-auth==0.1.2
django-pushserver==0.2.3
django-sekizai==0.5
django-staticfiles==1.2.1
-e git+https://github.com/mitar/django-tastypie.git@a1f2f95ece56aae7d265fb095dcefe3768a5a63c#egg=django-tastypie
django-tastypie-mongoengine==0.3
docutils==0.9.1
kombu==2.4.7
lxml==2.3
mimeparse==0.1.3
-e git+https://github.com/mitar/django-mongodbforms.git@01201c7b14127a3cb2cb027965320ed8ec499a4e#egg=mongodbforms-dev
mongoengine==0.7.5
-e git+https://github.com/mitar/django-mongogeneric.git@0d9d320398865d04e5c9d72b47543ecf68471813#egg=mongogeneric-dev
nose==1.2.1
-e git+https://github.com/mitar/py-hbpush.git@0.1.4-mitar#egg=py_hbpush-0.1.4
pymongo==2.3
python-dateutil==1.5
pytz==2012d
tornado==2.3
tweepy==1.9
wsgiref==0.1.2
